bifröst: uart_irq, uart_rdn, uart_wrn connected.
uart_rdn and uart_wrn are verified working.

uart_irq hasn't been verified yet, and relies on the pull-up specified
in the PCF file, which apparently is a thing that nextprn-ice40 and the
iCE40HX4K support, but I'll believe it when I see it. That said, it
hasn't triggered any spurious interrupts during a fair bit of testing,
so maybe it's working.